+++ CHANGES.rst
@@ -80,6 +80,9 @@ Bugs fixed
 * Compile time evaluations of (partially) constant f-strings could show incorrect
   results.
 
+* Escape sequences in raw f-strings (``fr'...'``) were resolved instead of passing
+  them through as expected.
+
 * Some ref-counting issues in buffer error handling have been resolved.
 
 Other changes